+++ modules/meteroid/Cargo.toml
@@ -69,6 +69,7 @@ rust_decimal_macros = { workspace = true }
 
 metering-grpc = { workspace = true, features = ["client"] }
 common-domain = { workspace = true }
+common-eventbus = { workspace = true }
 common-utils = { workspace = true, features = ["error-stack-conv"] }
 common-repository = { workspace = true }
 distributed-lock = { workspace = true, features = ["postgres-support"] }
