+++ modules/meteroid/crates/meteroid-store/Cargo.toml
@@ -9,6 +9,8 @@ repository.workspace = true
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+common-eventbus.workspace = true
+
 argon2 = { workspace = true }
 base62.workspace = true
 chrono.workspace = true
