+++ modules/meteroid/crates/meteroid-store/src/store.rs
@@ -1,8 +1,10 @@
 use crate::errors::StoreError;
 use diesel_async::pooled_connection::deadpool::Object;
 use diesel_async::pooled_connection::deadpool::Pool;
+use std::sync::Arc;
 
 use crate::StoreResult;
+use common_eventbus::{Event, EventBus};
 use diesel_async::pooled_connection::AsyncDieselConnectionManager;
 use diesel_async::scoped_futures::{ScopedBoxFuture, ScopedFutureExt};
 use diesel_async::{AsyncConnection, AsyncPgConnection};
@@ -15,6 +17,7 @@ pub type PgConn = Object<AsyncPgConnection>;
 pub struct Store {
     pub pool: PgPool,
     pub crypt_key: secrecy::SecretString,
+    pub eventbus: Arc<dyn EventBus<Event>>,
 }
 
 pub fn diesel_make_pg_pool(db_url: String) -> StoreResult<PgPool> {
@@ -29,10 +32,18 @@ pub fn diesel_make_pg_pool(db_url: String) -> StoreResult<PgPool> {
 }
 
 impl Store {
-    pub fn new(database_url: String, crypt_key: secrecy::SecretString) -> StoreResult<Self> {
+    pub fn new(
+        database_url: String,
+        crypt_key: secrecy::SecretString,
+        eventbus: Arc<dyn EventBus<Event>>,
+    ) -> StoreResult<Self> {
         let pool: PgPool = diesel_make_pg_pool(database_url)?;
 
-        Ok(Store { pool, crypt_key })
+        Ok(Store {
+            pool,
+            crypt_key,
+            eventbus,
+        })
     }
 
     pub async fn get_conn(&self) -> StoreResult<PgConn> {
