+++ modules/meteroid/src/api/apitokens/mod.rs
@@ -1,9 +1,5 @@
-use std::sync::Arc;
-
 use meteroid_grpc::meteroid::api::apitokens::v1::api_tokens_service_server::ApiTokensServiceServer;
 use meteroid_store::Store;
-
-use crate::eventbus::{Event, EventBus};
 
 mod error;
 mod mapping;
@@ -11,13 +7,9 @@ mod service;
 
 pub struct ApiTokensServiceComponents {
     pub store: Store,
-    pub eventbus: Arc<dyn EventBus<Event>>,
 }
 
-pub fn service(
-    store: Store,
-    eventbus: Arc<dyn EventBus<Event>>,
-) -> ApiTokensServiceServer<ApiTokensServiceComponents> {
-    let inner = ApiTokensServiceComponents { store, eventbus };
+pub fn service(store: Store) -> ApiTokensServiceServer<ApiTokensServiceComponents> {
+    let inner = ApiTokensServiceComponents { store };
     ApiTokensServiceServer::new(inner)
 }
