+++ modules/meteroid/src/bin/seeder.rs
@@ -1,9 +1,11 @@
 use std::env;
+use std::sync::Arc;
 
 use tokio::signal;
 
 use common_logging::init::init_regular_logging;
 use error_stack::ResultExt;
+use meteroid::eventbus;
 use meteroid::seeder::domain;
 use meteroid::seeder::errors::SeederError;
 use meteroid::seeder::runner;
@@ -25,6 +27,7 @@ async fn main() -> error_stack::Result<(), SeederError> {
     let store = Store::new(
         env::var("DATABASE_URL").change_context(SeederError::InitializationError)?,
         crypt_key,
+        Arc::new(eventbus::memory::InMemory::new()),
     )
     .change_context(SeederError::InitializationError)?;
