+++ modules/meteroid/src/eventbus/analytics_handler.rs
@@ -6,12 +6,11 @@ use uuid::Uuid;
 
 use common_build_info::BuildInfo;
 use common_config::analytics::AnalyticsConfig;
+use common_eventbus::{EventBusError, EventHandler};
 use common_logging::unwrapper::UnwrapLogger;
 use common_repository::Pool;
 
-use crate::eventbus::{
-    Event, EventBusError, EventData, EventDataDetails, EventHandler, TenantEventDataDetails,
-};
+use common_eventbus::{Event, EventData, EventDataDetails, TenantEventDataDetails};
 
 pub struct AnalyticsHandler {
     pool: Pool,
